try to fix windows build 4
