+++ cmake/components/qv2ray-lib.cmake
@@ -37,6 +37,7 @@ set(QV2RAY_LIB_SOURCES
     ${CMAKE_SOURCE_DIR}/src/components/latency/TCPing.cpp
     ${CMAKE_SOURCE_DIR}/src/components/latency/TCPing.hpp
     ${CMAKE_SOURCE_DIR}/src/components/latency/win/ICMPPing.cpp
+    ${CMAKE_SOURCE_DIR}/src/components/latency/win/ICMPPingWork.cpp
     ${CMAKE_SOURCE_DIR}/src/components/latency/win/ICMPPing.hpp
     ${CMAKE_SOURCE_DIR}/src/components/latency/unix/ICMPPing.cpp
     ${CMAKE_SOURCE_DIR}/src/components/latency/unix/ICMPPing.hpp
